requirements.txt
python-telegram-bot[webhooks]==20.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
requests==2.32.3
beautifulsoup4>=4.11.0
nest_asyncio
joblib
pydrive2
aiosqlite
aiohttp
PyGithub
python-dateutil
